+++ packages/react/src/components/Tooltip/index.tsx
@@ -25,14 +25,12 @@ export function Tooltip(props: TooltipProps) {
         <TooltipRdx.Trigger asChild>{props.tooltipTrigger}</TooltipRdx.Trigger>
         <TooltipRdx.Portal>
           <TooltipRdx.Content
-            className={`${tooltipClass} data-[state=delayed-open]:data-[side=top]:animate-slideDownAndFade data-[state=delayed-open]:data-[side=right]:animate-slideLeftAndFade data-[state=delayed-open]:data-[side=left]:animate-slideRightAndFade data-[state=delayed-open]:data-[side=bottom]:animate-slideUpAndFade text-violet11 select-none rounded-[4px] px-[15px] py-[10px] text-[15px] leading-none shadow-[hsl(206_22%_7%_/_35%)_0px_10px_38px_-10px,_hsl(206_22%_7%_/_20%)_0px_10px_20px_-15px] will-change-[transform,opacity]`}
+            className={`${tooltipClass} text-violet11 select-none rounded-[4px] px-[15px] py-[10px] text-[15px] leading-none shadow-[hsl(206_22%_7%_/_35%)_0px_10px_38px_-10px,_hsl(206_22%_7%_/_20%)_0px_10px_20px_-15px] will-change-[transform,opacity] data-[state=delayed-open]:data-[side=bottom]:animate-slideUpAndFade data-[state=delayed-open]:data-[side=left]:animate-slideRightAndFade data-[state=delayed-open]:data-[side=right]:animate-slideLeftAndFade data-[state=delayed-open]:data-[side=top]:animate-slideDownAndFade`}
             sideOffset={5}
           >
-            <Text
-              size={props.size}
-              content={props.textContent}
-              className={props.textClassName}
-            ></Text>
+            <Text size={props.size} className={props.textClassName}>
+              {props.textContent}
+            </Text>
             <TooltipRdx.Arrow className={`${arrowClass}`} />
           </TooltipRdx.Content>
         </TooltipRdx.Portal>